+++ src/extension.ts
@@ -7,6 +7,7 @@ import {
   createShowNormalizedViewFilteredByDateRangeCommand,
   createShowNormalizedViewFilteredByDateRangeAndSeverityCommand,
   createShowNormalizedViewFilteredByIgnorePatternCommand,
+  createShowNormalizedViewFilteredCommand,
   createShowCollapsedViewCommand,
 } from "./normalizedView";
 import {
@@ -65,6 +66,10 @@ export function activate(context: vscode.ExtensionContext): void {
     vscode.commands.registerCommand(
       "totonoeLog.showNormalizedViewFilteredByIgnorePattern",
       createShowNormalizedViewFilteredByIgnorePatternCommand(normalizedViewProvider)
+    ),
+    vscode.commands.registerCommand(
+      "totonoeLog.showNormalizedViewFiltered",
+      createShowNormalizedViewFilteredCommand(normalizedViewProvider)
     ),
     vscode.commands.registerCommand(
       "totonoeLog.showCollapsedView",